Add option to skip over parse failures.
Many dehacked files use non-vanilla features (like BEX extensions) and
it is helpful to be able to parse the whole of these files instead of
bombing out with an error half-way through the file. Add a strict_mode
parameter to allow the old behavior but default to false and print some
warnings at the end.