Simplify how 'original state' is represented.
We need to store the original state of all tables so that we can
compare modified against original state and generate a diff. But the
original approach compared against state at instantiation time and
stored a copy of structs in the constructor.

Instead, since we're always cloning the tables anyway whenever a
DehackedFile is instantiated, we only need to compare against the
original tables in tables.py that we're cloning from. Just set a
property named 'original' that gets overridden in the __copy__ methods.