Swap args/level order, accept string args.
It's convenient to be able to do dehfile.interactive("-foo -bar")
and have this work correctly. So make the args list the first argument
and use shlex to parse the value if it's a string rather than a tuple.
Also save the previous args to be consistent with the level arg.