Add sprite names list.
It's a common thing to want to change Doom's sprite names and this is
done by text replacements. However, it's also convenient to be able to
represent this as a sprites list like Doom has (the sprnames[] array).
Add a type that pretends to be a list but is really just a facade over
StringReplacements for convenience.