Only use relevant strategies to reclaim sprites.
Not all reclaim strategies cause sprites to be freed up, so it is
"wasteful" to run strategies which will degrade the game in other ways
while not achieving the set goal. To figure out which strategies are
relevant, run them first on a separate throw-away DehackedFile instance
to simulate the effect and see the result.