Change method for making animated items static.
Setting tics=-1 on a frame will cause the animation to halt on that
frame. However, this can also cause demo desyncs. Specifically, Doom's
P_SpawnMapThing() conditionally calls P_Random() if tics > 0, so
changing a thing to be static using this method will cause the RNG to
be called the wrong number of times.

Instead, make animations static by causing frames to point back to the
same frame again. This has the same effect and fixes the demo desyncs.