Fix desync when squashing MT_MISC75 animation.
As with previous commit 0e1e2dcacbb46fc8d, a change in initial state
from a frame with tics < 0 to tics >= 0 can cause a desync. So with the
twitching impaled body, instead of recovering both frames of animation
by reusing the non-twitching body frame, squash it down to just a single
frame. But we can still recover the sprite.