Assign sprites automatically during parsing.
This avoids an unnecessary call to assign_sprites() before parsing
and the possibility that the user may forget to pre-specify every
single sprite that is needed. This change also refactors the states
parser to be entirely self-contained; we separate the first, parsing
stage entirely from the second stage where we map the result into the
necessary tables.