reclaim: Squash pain elemental resurrect frames.
I assumed that the pain elemental resurrect frames were not used at
all, but scifista42 pointed out an obscure corner case where a pain
elemental can theoretically be crushed into gibs during its death
sequence and then resurrected by an archvile. It's so obscure as to
be practically ridiculous but isn't impossible. So retain one frame
for pain elemental resurrections and squash the resurrect animation
like we do for the other monsters, except that unlike the other
monsters the pain elemental's resurrect frames are the first to go.