Adjust plasma reload animation based on feedback.
Multiple people have criticized the original Perkristian plasma reload
animation; although "smooth" it's strange to look at (it's not clear
what the Doom marine is "doing" holding it up in the air).

This changes the animation to 1) use the original PLSGB sprite for the
reload period, and hold it for longer (15 frames, close to the original
20); 2) bring the weapon up into the reload position very quickly (only
two frames of animation); 3) retain smoother movement when bringing back
down to face forward again.

Thanks to Linguica for coming up with this in discussions on Doomworld.